implemented VulnerabilitiesController Create Method
