+++ WSafe/WSafe.Web/App_Start/BundleConfig.cs
@@ -4,7 +4,6 @@ namespace WSafe.Web
 {
     public class BundleConfig
     {
-        // For more information on bundling, visit https://go.microsoft.com/fwlink/?LinkId=301862
         public static void RegisterBundles(BundleCollection bundles)
         {
             bundles.Add(new ScriptBundle("~/bundles/jquery")
