+++ WSafe/WSafe.Web/Views/Shared/_EditAmenazas.cshtml
@@ -179,10 +179,6 @@
                         <input type="date" class="form-control" id="FechaFinal" required>
                     </div>
                 </div>
-
-            </div>
-
-
                 <div class="modal-footer">
                     <button type="button" class="btn btn-secondary" data-dismiss="modal">Cerrar</button>
                     <button type="button" class="btn btn-primary" onclick="addNewIntervention()">Guardar</button>
