+++ WSafe/WSafe.Web/Views/Vulnerabilities/Create.cshtml
@@ -41,7 +41,8 @@
             <div class="tabAmenaza">
                 <hr />
                 @Html.Partial("_EditAmenazas")
-                <input type="hidden" id="vulnerabilityID" name="vulnerabilityID"/>
+                <input type="hidden" id="vulnerabilityID" name="vulnerabilityID" />
+                <input type="hidden" id="interventionID" name="interventionID" />
             </div>
             <div style="display: inline-block;">
                 <div class="showIntervenciones"></div>
